Allow settings group_membership_claims on ad application
Set the groupMembershipClaims into AdditionalProperties[1]. This is an
unvalidated map that can be used to set arbitrary manifest entries.

This is kind of new for terraform however in this case the inputs are
well documented[2] and can be validated.

This imports the suppress case difference in strings helper from the
azuread provider.

[1] https://godoc.org/github.com/Azure/azure-sdk-for-go/services/graphrbac/1.6/graphrbac#ApplicationCreateParameters
[2] https://docs.microsoft.com/en-us/azure/active-directory/develop/reference-app-manifest