azuread_group - support for the description property (#216)
This is my first attempt at modifying code using this sdk and maybe I'm missing something but the sdk doesn't seem to expose the patch group call? Am I just blind? With this missing I've made groups recreate on modification on description but that's really kinda lame as an end user.

fixes #210