Stop requiring subscription ID to be configured for the provider
Configuring a subscription ID is a vestige from the provider split. We
don't use subscription_id anywhere and have no plans to. Any resource
that operates on a subscription or its dependents should belong in the
azurerm provider.

We'll keep the configuration property around for now so that users have
time to remove it from their configurations, and to avoid having to
remove the corresponding property from the `azuread_client_config` data
source. This will be dropped in v1.0.

Although it's never used because we only use SDK clients that are
configured with a tenant ID, go-azure-helper requires it for sensible
reasons and it doesn't make sense to push this concern upstream for now.
So for configurations where a subscription ID is not specified, we adopt
the tenant ID for the subscription ID. This mimics the Azure CLI which
does something similar when `--allow-no-subscriptions` is specified
(although we don't intend to expose this to users).