Fixes for some object ownership issues with applications, groups and service principals
Where no owners are specified in configuration, The calling principal
is explicitly assigned as the owner at creation time, and then removed
immediately afterwards. This provides stable behavior across resources
where sometimes the API will assign the caller as the owner and
sometimes not, and it's impossible to explicitly specify zero owners
at creation time.

Where owners are specified, 19 of them are assigned at creation time,
in addition to the calling principal. This is unordered since the
`owners` property is a TypeSet in all cases. The remaining owners (if
any) are subsequently assigned immediately after creation.

However for groups, the first 19 principals are preferentially
selected to be users, followed by service principals. This aims to
work around strict validation with Microsoft 365 groups where at least
one owner must be a user at all times! Ensuring that at least one user
is included in the set of owners in configuration is an exercise left
to the practitioner.

In all cases, if the calling principal is not explicitly included in the
set of owners in the configuration, it will not remain an owner
following an apply operation. This too is a decision left to the
practitioner to explicitly include this principal as an owner if so
required.

Create and Update timeouts are also increased for groups to 20 mins,
and applications & service princpals to 10 mins, to accommodate the time
needed to post large numbers of members/owners.