✨ add support for token signing certificate
We do like to add support to the provider to create a token signing certificate.

This commit adds new optional fields `token_signing_certificate_name` and
`token_signing_certificate_end_date` to the resource `azuread_service_principal`.
If both of these fields are set a token signing ceritificate for the
service principal gets created and attached.

The certificate thumbprint will be stored in a computed field called
`preferred_token_signing_key_thumbprint`.

There is now way to update or rotate the certificate because of missing
API support. Therefore in case of an adjustment the principal will
be replaced.

Issue: https://github.com/hashicorp/terraform-provider-azuread/issues/732