Add sentance documenting naming mismatch for apps and SPs
Every time I look into applications and service principals I have to double and tripple check the Azure documentation to see if I got them the right way round. Someone on the Azure discord just claimed it was the other way around so I guess I'm not the only one struggling with this.