implement creation of group directly in au via property
With this implementation it is possible to create a group directly in an
administrative unit provided by its id as property on a group resource.
There is currently a strange behavior where you've to wait for approx
10min until you can read the additional group property after creation
the group through the new implementation.
There is also no test written yet.